+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+Work in Progress
+----------------
+
+  * Fix fatal bug in failing to remove listeners when not using Express.
+
+
 v0.0.9 - 04/15/2013
 -------------------